ToyProject/requirements.txt
# Web Crawling
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
lxml>=4.9.0

# Data Storage
pandas>=2.0.0
pyarrow>=14.0.0
fastparquet>=2023.10.0

# LLM Support
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
pydantic>=2.5.0

